% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{HLA_A}
\alias{HLA_A}
\title{HLA-A}
\format{An object of class \code{\link[hlatools]{HLAGene}}.}
\source{
\href{https://www.ebi.ac.uk/ipd/}{IPD-IMGT/HLA}
}
\usage{
data(HLA_A)
}
\description{
HLA-A data from IPD-IMGT/HLA v. 3.25.0.
}
\examples{
data(HLA_A)
HLA_A$get_locusname()
HLA_A$get_db_version()
HLA_A$get_reference_sequence("01:01:01:01")
HLA_A$get_alleles("01:01:01:01")
}
\keyword{datasets}

