% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{HLA_B}
\alias{HLA_B}
\title{HLA-B}
\format{An object of class \code{\link[hlatools]{HLAGene}}.}
\source{
\href{https://www.ebi.ac.uk/ipd/}{IPD-IMGT/HLA}
}
\usage{
data(HLA_B)
}
\description{
HLA-B data from IPD-IMGT/HLA v. 3.25.0.
}
\examples{
data(HLA_B)
HLA_B$get_locusname()
HLA_B$get_db_version()
HLA_B$get_reference_sequence("07:02:02")
HLA_B$get_alleles("07:02:02")
}
\keyword{datasets}

