% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{HLA_C}
\alias{HLA_C}
\title{HLA-C}
\format{An object of class \code{\link[hlatools]{HLAGene}}.}
\source{
\href{https://www.ebi.ac.uk/ipd/}{IPD-IMGT/HLA}
}
\usage{
data(HLA_C)
}
\description{
HLA-C data from IPD-IMGT/HLA v. 3.25.0.
}
\examples{
data(HLA_C)
HLA_C$get_locusname()
HLA_C$get_db_version()
HLA_C$get_reference_sequence("01:02:01")
HLA_C$get_alleles("01:02:01")
}
\keyword{datasets}

