% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{HLA_DQB1}
\alias{HLA_DQB1}
\title{HLA-DQB1}
\format{An object of class \code{\link[hlatools]{HLAGene}}.}
\source{
\href{https://www.ebi.ac.uk/ipd/}{IPD-IMGT/HLA}
}
\usage{
data(HLA_DQB1)
}
\description{
HLA-DQB1 data from IPD-IMGT/HLA v. 3.25.0.
}
\examples{
data(HLA_DQB1)
HLA_DQB1$get_locusname()
HLA_DQB1$get_db_version()
HLA_DQB1$get_reference_sequence("05:01:01:01")
HLA_DQB1$get_alleles("05:01:01:01")
}
\keyword{datasets}

