% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{HLA_DRB1}
\alias{HLA_DRB1}
\title{HLA-DRB1}
\format{An object of class \code{\link[hlatools]{HLAGene}}.}
\source{
\href{https://www.ebi.ac.uk/ipd/}{IPD-IMGT/HLA}
}
\usage{
data(HLA_DRB1)
}
\description{
HLA-DRB1 data from IPD-IMGT/HLA v. 3.25.0.
}
\examples{
data(HLA_DRB1)
HLA_DRB1$get_locusname()
HLA_DRB1$get_db_version()
HLA_DRB1$get_reference_sequence("01:01:01")
HLA_DRB1$get_alleles("01:01:01")
}
\keyword{datasets}

