% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdata-package.r
\docType{package}
\name{IPDdata}
\alias{IPDdata}
\alias{IPDdata-package}
\title{IPDdata}
\description{
Allele data from IPD-IMGT/HLA
}
\details{
\tabular{ll}{
  Package: \tab IPDdata\cr
  Type: \tab Data package\cr
  Version: \tab 0.0.2\cr
  Date: \tab 05-01-2017\cr
  License: \tab MIT\cr
}
}
\author{
Gerhard Schöfl \email{schoefl@dkms-lab.de}
Maintainer: Gerhard Schöfl \email{schoefl@dkms-lab.de}
}

