% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{loadIPDdata}
\alias{loadIPDdata}
\title{Load and assign a data object from \pkg{IPDdata}.}
\usage{
loadIPDdata(locus)
}
\arguments{
\item{locus}{A valid HLA or KIR locus.}
}
\value{
A \code{\link[hlatools]{HLAGene}} object
}
\description{
Load and assign a data object from \pkg{IPDdata}.
}
\examples{
x <- loadIPDdata("A")
}

